test(e2e): web: extract Event Logs ACL helpers for reuse
Address review feedback: make the ACL setup and Event Logs page
helpers reusable by other tests.

- move the CLAPI ACL-profile loader to a global `applyAclProfile`
  command in support/commands.ts;
- move the Event Logs page helpers to features/Logs/common.ts.

Refs: MON-159125
